Fix ordering of OriginalFile objects registered in saveFileset()
During the import process, the server receives a Fileset object and a
matching list of paths. The saveFileset() API first registers these
paths as OriginalFile then associate these files to each fileset entry
in order.

The current implementation of _internalRegister groups the input checked
paths by parent using a ListMultimap and iterates through each group to
create or retrieve the OriginalFile objects. Under some conditions, the
returned list will be in a different order than the input list.

This commit does not modify the mapping and creation logic but adds a
final step to return a list of OriginalFile sorted in the same order
as the input checked paths. This contract is also added to the Javadoc.