ManagedImportProcessI.verifyUpload() improvements for large filesets
The currently implementation of the checksum validation retrieves the
server hashes by executing as many HQL queries as the number of
OriginalFile objects in the fileset.
For filesets containing 1K-100K files which is fairly common in the
high-content screening domains and individual queries taking ~100-200ms,
this can lead to multiple hours of checksum verification.

This commit updates the logic to use a single HQL query and create a
hash map of all serverHashes for the fileset indexed by their originalfile
path/name. This map is then used in the following loop to compare each
hash to the client checksum.