OMEROMetadataStoreClient: fall back to secure session on SocketException
Following authentication using the 4064 port (SSL), the importer will
attempt to create an unsecure client using TCP port 4063.
Currently if this session creation results in an
Ice.ConnectionRefusedException exception e.g. because port 4063 is
blocked, it will fall back to the secure client and continue the
import.

Under some conditions that might depend both on the operating system
and Java, `unsecure()` might throw an
`Ice.SocketException: java.net.NoRouteToHostException: No route to host`
which currently aborts the import.

This commit proposes to relax the exception handling to catch any
Ice.SocketException and fallback to a secure connection.